created Piece, and King. King and Man inherit from Piece now
+ start of King movement (mainly Pseudo code)
