+++ Assets/Scripts/GameObjects/Controller.cs
@@ -70,96 +70,13 @@ public class Controller : MonoBehaviour
 
         selected = tileObject; 
 
-        Man man = tileObject as Man;
-        man.SetAvailableMoves(CheckAvailableMoves(man.IsMovingRight, man));
+        Piece piece = tileObject as Piece;
+        piece.SetAvailableMoves(piece.CheckAvailableMoves());
         
-        foreach(var t in man.AvailableMoves)
+        foreach(var t in piece.AvailableMoves)
         {
             Debug.Log(t.ToString());
         }
     }
 
-    List<Tile> CheckAvailableMoves(bool isMovingRight, Man manToCheck)
-    {
-        List<Tile> availablePositions = new List<Tile>();
-        if (!isMovingRight)
-        {
-            // check NE/SE
-            Tile? NETile = CheckPointForward(Point.NorthEast);
-
-            //if not null then add to availablePositions
-            if (NETile != null)
-            {
-                availablePositions.Add(NETile);
-            }
-
-            Tile? SETile = CheckPointForward(Point.SouthEast);
-
-            if (SETile != null)
-            {
-                availablePositions.Add(SETile);
-            }
-        }
-        else
-        {
-            // check NW/SW
-            Tile? NWTile = CheckPointForward(Point.NorthWest);
-
-            //if not null then add to availablePositions
-            if (NWTile != null)
-            {
-                availablePositions.Add(NWTile);
-            }
-
-            Tile? SWTile = CheckPointForward(Point.SouthWest);
-
-            if (SWTile != null)
-            {
-                availablePositions.Add(SWTile);
-            }
-        }
-
-        return availablePositions;
-    }
-
-
-    Tile CheckPointForward(Point direction)
-    {
-        Tile? tileInDirection = Core.Main.Tilemap.GetTile(selected.Tile.Position + direction);
-
-        if (tileInDirection == null)
-        {
-            return null;
-        }
-
-        TileObject? tileObjectInDirection = tileInDirection.tileObject;
-
-        if (tileObjectInDirection == null)
-        {
-            return tileInDirection;
-        }
-        //if ally tileobject break
-        if (tileObjectInDirection.Actor == Core.Main.TurnManager.CurrentTurn)
-        {
-            return null;
-        }
-
-        //if enemyObject check another NE/SE
-        Tile? tileInDirectionAgain = Core.Main.Tilemap.GetTile(tileInDirection.Position + direction);
-
-        if (tileInDirectionAgain == null)
-        {
-            return null;
-        }
-
-        TileObject? tileObjectInDirectionAgain = tileInDirectionAgain.tileObject;
-
-        if (tileObjectInDirectionAgain != null)
-        {
-            return null;
-        }
-
-        return tileInDirectionAgain;
-
-    }
 }
