+++ Assets/Scripts/GameObjects/King.cs
@@ -1,0 +1,28 @@
+﻿using Mishbetzet;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Assets.Scripts.GameObjects
+{
+    public class King : Piece
+    {
+        public override List<Tile> CheckAvailableMoves()
+        {
+            //CheckPointForward
+
+            //continue in diagonal 8 moves
+            //if tile is empty return tile
+            //if tile has object, check if object behind
+            //if object behind it then return
+            //if no object then return tile
+            
+            //while loop for every direction, with nested if for double object
+
+
+            return null;
+        }
+    }
+}
