+++ Assets/Scripts/GameObjects/King.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d25ddd3bd3eb27418187192a46940b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
