+++ Assets/Scripts/GameObjects/Piece.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f59ff514b2ae9b148b9d18a001333d89
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
